Created new gameobjectvariable to track a current "building" item, which will track the mouse position when set. also implemented custom property drawer for the new BooleanReference value, which can switch between a constant and a variable boolean.
Added new tile member management pattern which tracks all members by their coordinates, for easier indexing when looking for contents of a tile. used this pattern to detect when a tile is passable based on the members in that tile in addition to the tilemap properties.

Allow for tilemapMembers to switch to different tilemapRegions