moved time into a scriptable object
refactored all states into generic states applied to any TileMapMember. Created new task scriptable object which can be composed and used to build new supertasks. Created a scheduler which will choose between a list of tasks in priority order

created pseudo-chainable contineuableState. used it in combination with a simple Delay state to replace the functionality of the now-defunct Waiting state