Split building and supplying for building into two different tasks
Created new inventory variable, to allow members to keep track of multiple different inventories. This is useful to seperate items used for building from items produced by a plant, or items consumed by the machine.

Created new Supplyable component, which leverages the new Inventory variable. represents an ability for an actor to supply that inventory variable with something. Currently it will flip a boolean variable when its inventory is full and can recieve no more items

Removed the inventory responsibility from Buildable, made it a more generic component handle which just hooks up to boolean variables. The variables will be updated by other components, such as a Supplyable which can flip a value when it is full and ready for building