overhauled the data input paradigm for the connectivity system, was able to get the startup for the job down to 10-15ms even for a 200x200 tilemap.
pass in raw tile data into the job, and calculate passability inside the job instead of in the frame when launching. store tile data as a nativeHashMap now instead of as a dictionary, which allows passing it directly into a job.

No longer build up an array of graph members, instead iterate over the members and assign regionIDs on job completion.

build a nativeArray of coordinate data from the beginning instead of using a list. the list generated a very large amount of garbage, allocating one known-size nativeArray cuts down on this significantly

Further optimization would involve passing only the coordinate range data and sets of disabled coordinates directly, instead of populating that native array on the main thread